Renamed decoders to solvers
Since the base class is Solver.
