+++ nengo/builder.py
@@ -39,7 +39,6 @@ import warnings
 
 import numpy as np
 
-import nengo.decoders
 import nengo.neurons
 import nengo.objects
 import nengo.synapses
