+++ nengo/objects.py
@@ -6,11 +6,11 @@ import warnings
 
 import numpy as np
 
-import nengo.decoders
 from nengo import params
 from nengo.config import Config
 from nengo.neurons import LIF
 from nengo.params import Default, is_param
+from nengo.solvers import LstsqL2
 from nengo.synapses import Lowpass
 from nengo.utils.compat import with_metaclass
 from nengo.utils.distributions import Uniform, UniformHypersphere
@@ -641,7 +641,7 @@ class Connection(NengoObject):
     post = params.NengoObjectParam(disallow=[])
     synapse = params.SynapseParam(default=Lowpass(0.005))
     transform = params.TransformParam(default=np.array(1.0))
-    solver = params.SolverParam(default=nengo.decoders.LstsqL2())
+    solver = params.SolverParam(default=LstsqL2())
     function_info = params.FunctionParam(default=None, optional=True)
     modulatory = params.BoolParam(default=False)
     learning_rule = params.LearningRuleParam(default=None, optional=True)
