+++ nengo/solvers.py
